% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning-helpers.R
\name{localize_beach}
\alias{localize_beach}
\title{Add counrty to beach}
\usage{
localize_beach(dat)
}
\arguments{
\item{dat}{A table.}

\item{lookup_table}{B table.}
}
\value{
A left join B returning new table \code{A}
}
\description{
Add counrty to beach
}
